+++ src/main/java/roomescape/controller/ReservationController.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import roomescape.dto.Reservation;
-import roomescape.exception.NotFoundReservationException;
+import roomescape.dto.exception.NotFoundReservationException;
 
 import java.net.URI;
 import java.util.ArrayList;
