+++ src/main/java/roomescape/dto/exception/NotFoundReservationException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.dto.exception;
 
 public class NotFoundReservationException extends RuntimeException {
     public NotFoundReservationException(String message) {
